Cleaned up some code, modulated types for site data.
