+++ src/assets/siteDataTypes.ts
@@ -1,0 +1,13 @@
+export type NavLink = {
+  name: string,
+  link: string
+}
+
+export type Project = {
+  name: string,
+  stack: string[],
+  description: string,
+  link: string,
+  image: string,
+  gif?: string,
+}
