+++ src/components/project-card/ProjectBtns.tsx
@@ -6,12 +6,11 @@ interface Props {
   link: string;
 }
 
-const ProjectBtns = ({ link, handleSetActive }:Props) => {
-  return (
-    <div className="flex gap-2 absolute right-4 bottom-4">
-      <FlipBtn Flip={handleSetActive} />
-      <GithubBtn link={link} />
-    </div>
-  )
-}
+const ProjectBtns = ({ link, handleSetActive }:Props) => (
+  <div className="flex gap-2 absolute right-4 bottom-4">
+    <FlipBtn Flip={handleSetActive} />
+    <GithubBtn link={link} />
+  </div>
+)
+
 export default ProjectBtns
