+++ src/components/project-card/UsedStack.tsx
@@ -3,15 +3,13 @@ interface Props {
   skills: string[];
 }
 
-const UsedStack = ({ projectName, skills }:Props) => {
-  return (
-    <div className="flex gap-2 items-center justify-center">
-      {
-        skills.map(skill => {
-          return <i key={`${projectName}-${skill}`} className={`devicon-${skill} colored text-3xl`}></i>
-        })
-      }
-    </div>
-  )
-}
+const UsedStack = ({ projectName, skills }:Props) => (
+  <div className="flex gap-2 items-center justify-center">
+    {
+      skills.map(skill => {
+        return <i key={`${projectName}-${skill}`} className={`devicon-${skill} colored text-3xl`}></i>
+      })
+    }
+  </div>
+)
 export default UsedStack;
